add no quant variant on fused_clamp_act_mul
